+++ apps/prairielearn/src/tests/permissions/studentData.test.js
@@ -6,6 +6,7 @@ const sqldb = require('@prairielearn/postgres');
 const sql = sqldb.loadSqlEquiv(__filename);
 const helperServer = require('../helperServer');
 const helperClient = require('../helperClient');
+const { ensureEnrollment } = require('../../models/enrollment');
 
 describe('student data access', function () {
   this.timeout(60000);
@@ -46,7 +47,7 @@ describe('student data access', function () {
       'Owner',
       1,
     ]);
-    await sqldb.queryAsync(sql.insert_enrollment, {
+    await ensureEnrollment({
       user_id: 3,
       course_instance_id: 1,
     });