+++ packages/docker-utils/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @prairielearn/docker-utils
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [6cad75197]
+  - @prairielearn/logger@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [10cc07dcc]
+  - @prairielearn/logger@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [098f581da]
+  - @prairielearn/logger@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes