+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/opentelemetry
 
+## 1.8.6
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+
+## 1.8.5
+
+### Patch Changes
+
+- 10cc07dcc: Upgrade all dependencies
+
+## 1.8.4
+
+### Patch Changes
+
+- 098f581da: Upgrade all dependencies
+
 ## 1.8.3
 
 ### Patch Changes