+++ packages/sentry/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/sentry
 
+## 1.0.8
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+
+## 1.0.7
+
+### Patch Changes
+
+- 10cc07dcc: Upgrade all dependencies
+
+## 1.0.6
+
+### Patch Changes
+
+- 098f581da: Upgrade all dependencies
+
 ## 1.0.5
 
 ### Patch Changes